First working version of scene parser
see: http://tinyurl.com/cemqhqe

All elements are correctly created now, but there is still
a lot of things which need to be handeled:

a) Movement is now realized with a fixed value for the hero, this should
be dynamic.

b) When resizing the browser viewport, the object maintain their
position but do not scale accordingly to new viewport size.

c) The movement animation created by Vendetta 1987 should be integrated
into the scene creation.

d) There are still some issues when the hero hits the viewports borders.
Sometimes there sticks to the border.