Directional walking animations - Initial version.
See: http://tinyurl.com/bznroe6

I switched from 1 dimensional tilesets to
2 dimensional tilesets only for person objects.
This makes it quite easy to change the walk animation,
depending on the current directional vector.

However this comes not with a downside. There are now some
scaling issues, due to the increased tileset height,
which need to be fixed. In addition to the missing border collision,
this leads in some cases to wrong object movement.