Fix for buggy image clipping.
Initial division by tileset number per person,
is now done only after the first object has been
clicked, and not every time an object has
been clicked.

This is just a workaround. The more suitable version
creates the canvas in the right size and doesn't change
it's size when movement is initiated.