New clickEventHandler.js
I added a new file for a function in which all
actions (walk animation, quiz proceeding, dialog triggers)
are invoked. This is basically the same like before, when
all functions were called directly in the onclick:"" event.

Right now it just makes the code a bit more neat and readable. My long
term goal for this eventhandler is, to get a better control over the
sucession of events and an effective way to avoid event interference
between the single sub-functions. This will be done by additional logic
which allows sub-functions to be carried out or not depending on the
state of the game.

For instance, I would like to see it, that when a new quiz step is
reached, the corresponding dialogs have to be clicked through first,
before the next quiz step can be invoked. Same with walk animation.
This should be finished before a dialog/quizstep is triggered.