Added some additional logic to clickEventHandler.js
bewegePerson(), advanceQuizStep(), advanceDialogStep()
will now be activated in correct order. While the person
moves, quiz and dialog proceeding is deactivated. The player has to
click through dialog a chunk in the dialog box, before the next step
will be invoked. Clicking on an object other than the dialogbox, while
in the middle of a textchunk, will only proceed the dialog, but not the
games state (only exception to that is, when dialog activates the next
step).

These changes work for me so far, but it needs testing, testing,
testing!