Further work on scene 2 and a new functionality for dialog based quiz step activation.
It is now possible to activate the next quizstep at start of a dialog
chunk or at it's end. For this the new flag "trigger_at_start" has to be
set 'true' or 'false' in dialog.xml. Only if increase_quiz_step is
'true' this flag will be considered.
This was necessary because some events in scene 2 have to be triggered
head up start while others at the end of a dialog chunk.

Scene 2 is almost done. There is still a minor glitch at the end of the
scene. The initial automatic dialog activation has to be redone as well.