New functionality in dialogControl.js...
...and further work on scene 5.

I found it was necessary to manage dialogs with IDs like X.Y.Z in a
better way. In scene 4 those dialogs weren't depicted like it was
intended.

Those dialogues are now handled differently in contrast to the rest of
the dialogues. Still, this is experimental and only tested for scene 4.
So some feedback would be in order.

Moreover I did some changes to scene 5, but it's still far from being
done.