Prototype for an scene exception handler
We have dire need of a mechanism, which handles special animations
and riddles for specific scenes. I implemented a prototype for
this case. Basically it is now possible to trigger out of order
activities, simply by passing two attribute values in dialoge.xml
to triggerException(). As an example tested it for the book riddle in
scene 5. So far it works well. For a more thorough explanation I refer
you to the comment block at the head of dialoge.xml.