New Exception for the big blow out finale
I used a combination of transformation steps
to create a neat little animation of dr chaos
being sucked out of the freight hangar, when the
freight door is openend. To make it a bit more
"chaotic" the position is calculated after the
formula for an archimedean spiral. This looks
better than just rotating the image.