Prototype for a flipping function
Another exception for scene 7. Spike has
to turn around and check if the door is
really locked like dr chaos said. So this will
flip spike or any other character, whose gbilder
id has been passed, back and couple of seconds
back to it's original state. Right now the image
is not properly scaled, so this remains to be done.