Fix for scaling bugs in exceptions
An frequently occuring task in exceptions
is to redraw an image differently from its
start adjustment. Applied transformations
require it to clear the canvas an redraw it
with the same width height values which were
used by the scene parser to render the scene.
For this gImageStats was introduced an is now
utilized by getScaledDimensions() in sceneExceptions.js.
This allows it to redraw the images with the correct
scaling.