experimental WalkAnimation scaling fix
wrong scaling is due to gStartAbmessungen not reaching its projected
value because of the propably to few steps while scaling on the central
path. seting it to the projected value by force fixes it for now but i'm
going to look into it again to make it scale correctly while on the
central path instead.