DebugMode
Added a debug mode to display every function call in Firebug. It can be
turned on by switching gDebugModeOn on helper.js. I also changed the
code loading to just set the scene number for the moment. This way we
always start at the beginning for the time this isn't properly fixed.