Dialogue bug
I further investigated the problem and it seems to be related to this
function. the single one change i did to dialogcontrol while
implementing loading by code was to replace the two break statements by
one return. this shouldn't change anything and i reverted it to test. it
works as intended. read the comment for further information. the problem
is that fetchdialogids returns he wrong array at the end of scene 1.