Bugfix for Scene 4
Lock happened, when not all sub dialogues were triggered in advance
before clicking on the oil can the second time. As a result the
sub dialogues weren't blacklisted and messed up the sub dialog
offset counter. This is fixed now by adding all sub dialogues
of the current sub dialog group to the blacklist, when the last
dialog of that group is triggered.